third_party/WebKit/LayoutTests/fast/speech/intrinsic-input-width-with-speech-border-padding-expected.txt

CONSOLE WARNING: The 'x-webkit-speech' input field attribute is deprecated. Please use the JavaScript API instead.
PASS document.getElementById("speechInput1").offsetWidth is document.getElementById("input1").offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
 
